# [quantization] Remove index calculation in quantized max_pool2d